Fix label type issue

+++ examples/ssgan_tutorial.py
@@ -202,7 +202,7 @@ class SSGANDiscriminator(gan.BaseDiscriminator):
         # Compute SS loss
         _, output_classes = self.forward(images_rot)
 
-        err_SS = F.cross_entropy(input=output_classes, target=class_labels)
+        err_SS = F.cross_entropy(input=output_classes, target=class_labels.long())
 
         # Scale SS loss
         err_SS = scale * err_SS
